refactor: extract SPLIT + verify survivors from the account tree
Phase 3 of the CFS to hiveid account move: restructure so CFS keeps serving the
SPLIT tournament routes + verify infra once the MOVE AccountModule is dropped at
the nginx cutover. Behavior-preserving (coexistence) and boot-verified.

New survivor modules:
- TournamentAuthModule owns the SPLIT relay-token mints (TournamentTokenController
  + TrackerTokenService + SplitTokenSigner), the SPLIT hiveid-tournament routes
  (HiveIDTournamentController + HiveIDTournamentService, extracted from the MOVE
  HiveIDService), the global AuthGuard, the JWT-verifying AuthMiddleware, and the
  JWKS controller. tracker-token / jwks / the token+claim DTOs physically moved
  here (git mv); decorators/guards/middleware stay in account/auth as allowlisted
  verify-infra exceptions.
- TournamentAdminModule re-parents the STAY registrations + declarations surface.

AuthMiddleware decoupled from the MOVE AuthService: it now verifies via the neutral
verifyJwt, so verify infra no longer depends on the account tree. The 3 SPLIT token
routes are removed from AuthController and the 3 tournament routes from
HiveIDController (served only by the survivors — no collision).

AppModule loads the survivors before AccountModule. The account boundary allowlist
gains the auth.guard / auth.middleware verify exceptions + the re-parented STAY
module entrypoints. JwtModule.register stays in AuthModule for coexistence (drop-
step: move it to TournamentAuthModule when AccountModule is removed).

Live boot verified on :8390 (Postgres+Redis): SPLIT routes + JWKS + MOVE routes all
served (401/400/200, no 404), bogus route 404, no DI/route-collision errors. Full
suite 1261 green; lint + account-boundary clean.
