+++ src/modules/account/auth/hiveid.service.ts
@@ -31,11 +31,8 @@ import {
   type IUserStorage,
 } from 'src/storage/interfaces';
 import { ConfigService } from '@nestjs/config';
-import { CANONICAL_PERSON, HIVEID_MAGIC_LINK_PREFIX } from './hiveid.constants';
-import { executionQueue as runExecutionQueue } from '../../factory/functions/private/executionQueue';
+import { HIVEID_MAGIC_LINK_PREFIX } from './hiveid.constants';
 import { PersonsClient, type PersonFragmentInput } from '../persons/persons-client.service';
-import { TournamentStorageService } from 'src/storage/tournament-storage.service';
-import { AuditService } from '../../audit/audit.service';
 import type { HiveIDSignupDto } from './dto/hiveidSignup.dto';
 import { IdentityService } from '../identity/identity.service';
 import { EmailService } from '../email/email.service';
@@ -45,58 +42,10 @@ import { AuthService } from './auth.service';
 const HIVEID_MAGIC_LINK_TTL_MINUTES = 15;
 const HIVEID_MAGIC_LINK_TTL_MS = HIVEID_MAGIC_LINK_TTL_MINUTES * 60 * 1000;
 
-export interface ParticipationRow {
-  tournamentId: string;
-  tournamentName: string;
-  startDate: string | null;
-  endDate: string | null;
-  participantId: string;
-  participantName: string;
-  eventCount: number;
-}
-
-export interface ClaimableCandidate {
-  participantId: string;
-  participantName: string;
-  sex: string | null;
-  nationalityCode: string | null;
-  birthDate: string | null;
-  alreadyLinkedTo: string | null;
-}
-
-function normalizeName(value: string): string {
-  return String(value ?? '').trim().toLowerCase();
-}
-
-function isIndividualParticipant(participant: any): boolean {
-  return (participant?.participantType ?? 'INDIVIDUAL') === 'INDIVIDUAL';
-}
-
-function extractCanonicalPersonId(participant: any): string | null {
-  const otherIds: any[] = participant?.person?.personOtherIds ?? [];
-  const hit = otherIds.find((o) => o?.organisationId === CANONICAL_PERSON);
-  return hit?.personId ?? null;
-}
-
-function participantMatchesPerson(participant: any, personId: string): boolean {
-  return extractCanonicalPersonId(participant) === personId;
-}
-
-function countParticipantEvents(tournament: any, participantId: string): number {
-  let count = 0;
-  for (const event of tournament?.events ?? []) {
-    const entries = event?.entries ?? [];
-    if (entries.some((e: any) => e?.participantId === participantId)) count++;
-  }
-  return count;
-}
-
-function byStartDateDesc(a: ParticipationRow, b: ParticipationRow): number {
-  const aDate = a.startDate ?? '';
-  const bDate = b.startDate ?? '';
-  if (aDate === bDate) return a.tournamentName.localeCompare(b.tournamentName);
-  return aDate < bDate ? 1 : -1;
-}
+// The SPLIT tournament methods (getMyParticipations / getClaimableForTournament /
+// claimParticipant) plus their participant helpers + row/candidate types moved to
+// HiveIDTournamentService (tournament-auth) so they survive the Phase-3 drop of
+// this MOVE service — they read CFS tournament records and stay on CFS.
 
 @Injectable()
 export class HiveIDService {
@@ -107,8 +56,6 @@ export class HiveIDService {
     private readonly identityService: IdentityService,
     private readonly configService: ConfigService,
     private readonly personsClient: PersonsClient,
-    private readonly tournamentStorageService: TournamentStorageService,
-    private readonly auditService: AuditService,
     @Inject(USER_STORAGE) private readonly userStorage: IUserStorage,
     @Inject(AUTH_CODE_STORAGE) private readonly authCodeStorage: IAuthCodeStorage,
   ) {}
@@ -511,175 +458,6 @@ export class HiveIDService {
     };
   }
 
-  /**
-   * GET /auth/hiveid/me/participations — surface all tournaments where
-   * the caller has been claimed as a Participant via the CANONICAL_PERSON
-   * organisationId. Scans every tournament record server-side; acceptable
-   * for early adoption (handful of HiveID users) and trivially cacheable
-   * later when volume grows.
-   */
-  async getMyParticipations(userId: string): Promise<{
-    personId: string | null;
-    participations: ParticipationRow[];
-  }> {
-    if (!userId) throw new UnauthorizedException();
-    const link = await this.userStorage.getPersonLink(userId);
-    const personId = link?.personId ?? null;
-    if (!personId) return { personId: null, participations: [] };
-
-    const tournamentIds = await this.tournamentStorageService.listTournamentIds();
-    if (!tournamentIds.length) return { personId, participations: [] };
-
-    const { tournamentRecords } = (await this.tournamentStorageService.fetchTournamentRecords({
-      tournamentIds,
-    })) as { tournamentRecords?: Record<string, any> };
-    if (!tournamentRecords) return { personId, participations: [] };
-
-    const out: ParticipationRow[] = [];
-    for (const tournament of Object.values(tournamentRecords)) {
-      const tid = tournament?.tournamentId;
-      if (!tid) continue;
-      const participants: any[] = tournament?.participants ?? [];
-      for (const participant of participants) {
-        if (!participantMatchesPerson(participant, personId)) continue;
-        out.push({
-          tournamentId: tid,
-          tournamentName: tournament.tournamentName ?? '',
-          startDate: tournament.startDate ?? null,
-          endDate: tournament.endDate ?? null,
-          participantId: participant.participantId,
-          participantName: participant.participantName ?? '',
-          eventCount: countParticipantEvents(tournament, participant.participantId),
-        });
-      }
-    }
-    out.sort(byStartDateDesc);
-    return { personId, participations: out };
-  }
-
-  /**
-   * GET /auth/hiveid/me/claimable/:tournamentId — returns the Participants
-   * in the given tournament whose canonical name overlaps the caller's
-   * cached fields, MINUS anyone already linked to the caller's personId.
-   * Defense-in-depth: the actual claim mutation re-verifies before
-   * stamping (see `claimParticipant`).
-   */
-  async getClaimableForTournament(userId: string, tournamentId: string): Promise<{
-    tournamentId: string;
-    candidates: ClaimableCandidate[];
-  }> {
-    if (!userId) throw new UnauthorizedException();
-    if (!tournamentId) throw new BadRequestException('tournamentId is required');
-
-    const link = await this.userStorage.getPersonLink(userId);
-    const personId = link?.personId ?? null;
-    const cached = link?.cached;
-    if (!cached?.standardGivenName || !cached?.standardFamilyName) {
-      return { tournamentId, candidates: [] };
-    }
-
-    const { tournamentRecord } = await this.tournamentStorageService.findTournamentRecord({ tournamentId });
-    if (!tournamentRecord) return { tournamentId, candidates: [] };
-
-    const targetGiven = normalizeName(cached.standardGivenName);
-    const targetFamily = normalizeName(cached.standardFamilyName);
-    const participants: any[] = tournamentRecord.participants ?? [];
-    const candidates: ClaimableCandidate[] = [];
-    for (const p of participants) {
-      if (!isIndividualParticipant(p)) continue;
-      if (personId && participantMatchesPerson(p, personId)) continue;
-      const personGiven = normalizeName(p?.person?.standardGivenName ?? p?.person?.givenName ?? '');
-      const personFamily = normalizeName(p?.person?.standardFamilyName ?? p?.person?.familyName ?? '');
-      const nameMatches =
-        (personGiven && personGiven === targetGiven) ||
-        (personFamily && personFamily === targetFamily);
-      if (!nameMatches) continue;
-      candidates.push({
-        participantId: p.participantId,
-        participantName: p.participantName ?? '',
-        sex: p?.person?.sex ?? null,
-        nationalityCode: p?.person?.nationalityCode ?? null,
-        birthDate: p?.person?.birthDate ?? null,
-        alreadyLinkedTo: extractCanonicalPersonId(p),
-      });
-    }
-    return { tournamentId, candidates };
-  }
-
-  /**
-   * POST /auth/hiveid/me/claim — stamp a `CANONICAL_PERSON`-keyed entry
-   * onto the target Participant's `Person.personOtherIds[]` via the
-   * `addPersonOtherId` factory mutation (PR-K). Defense-in-depth: the
-   * server reloads the tournament and re-verifies the participant exists
-   * and that the cached name overlaps before firing.
-   */
-  async claimParticipant(args: {
-    userId: string;
-    tournamentId: string;
-    participantId: string;
-    auditSource?: string;
-  }): Promise<{ success: true; tournamentId: string; participantId: string; personId: string }> {
-    const { userId, tournamentId, participantId } = args;
-    if (!userId) throw new UnauthorizedException();
-    if (!tournamentId || !participantId) {
-      throw new BadRequestException('tournamentId and participantId are required');
-    }
-
-    const link = await this.userStorage.getPersonLink(userId);
-    const personId = link?.personId;
-    if (!personId) {
-      throw new BadRequestException('Your HiveID does not yet have a canonical link.');
-    }
-
-    const { tournamentRecord } = await this.tournamentStorageService.findTournamentRecord({ tournamentId });
-    if (!tournamentRecord) throw new BadRequestException('Tournament not found');
-    const target = (tournamentRecord.participants ?? []).find((p: any) => p?.participantId === participantId);
-    if (!target) throw new BadRequestException('Participant not found in tournament');
-    if (!isIndividualParticipant(target)) {
-      throw new BadRequestException('Only INDIVIDUAL participants can be claimed');
-    }
-
-    const cached = link.cached;
-    const targetGiven = normalizeName(cached?.standardGivenName ?? '');
-    const targetFamily = normalizeName(cached?.standardFamilyName ?? '');
-    const personGiven = normalizeName(target?.person?.standardGivenName ?? target?.person?.givenName ?? '');
-    const personFamily = normalizeName(target?.person?.standardFamilyName ?? target?.person?.familyName ?? '');
-    const overlap =
-      (targetGiven && targetGiven === personGiven) || (targetFamily && targetFamily === personFamily);
-    if (!overlap) {
-      throw new BadRequestException(
-        'The target participant does not match your canonical name. If this is you, contact the tournament director.',
-      );
-    }
-
-    const result = await runExecutionQueue(
-      {
-        tournamentIds: [tournamentId],
-        methods: [
-          {
-            method: 'addPersonOtherId',
-            params: {
-              tournamentId,
-              participantId,
-              organisationId: CANONICAL_PERSON,
-              personId,
-            },
-          },
-        ],
-        userId,
-        userEmail: undefined,
-        source: args.auditSource ?? 'hiveid-claim',
-      },
-      undefined,
-      this.tournamentStorageService,
-      this.auditService,
-    );
-    if (!result?.success) {
-      const error = result?.error ?? 'addPersonOtherId mutation failed';
-      throw new BadRequestException(typeof error === 'string' ? error : JSON.stringify(error));
-    }
-    return { success: true, tournamentId, participantId, personId };
-  }
 
   /**
    * Magic-link URL surface for HiveID lands on the courthive-public app,
