+++ src/modules/tournament-auth/dto/hiveidClaim.dto.ts
