+++ src/modules/tournament-auth/dto/providerScoringToken.dto.ts
