+++ src/modules/tournament-auth/dto/scorerToken.dto.ts
